:tractor: Move client into go-go-mcp
