+++ Makefile
@@ -35,14 +35,14 @@ tag-patch:
 
 release:
 	git push --tags
-	GOPROXY=proxy.golang.org go list -m github.com/go-go-golems/XXX@$(shell svu current)
+	GOPROXY=proxy.golang.org go list -m github.com/go-go-golems/go-go-mcp/cmd/go-go-mcp@$(shell svu current)
 
 bump-glazed:
 	go get github.com/go-go-golems/glazed@latest
 	go get github.com/go-go-golems/clay@latest
 	go mod tidy
 
-XXX_BINARY=$(shell which XXX)
+go-go-mcp_BINARY=$(shell which go-go-mcp)
 install:
-	go build -o ./dist/XXX ./cmd/XXX && \
-		cp ./dist/XXX $(XXX_BINARY)
+	go build -o ./dist/go-go-mcp ./cmd/go-go-mcp && \
+		cp ./dist/go-go-mcp $(go-go-mcp_BINARY)
