+++ cmd/go-go-mcp/cmds/client/tools.go
@@ -13,8 +13,8 @@ import (
 	"github.com/go-go-golems/glazed/pkg/middlewares"
 	"github.com/go-go-golems/glazed/pkg/settings"
 	"github.com/go-go-golems/glazed/pkg/types"
-	"github.com/go-go-golems/go-go-mcp/cmd/mcp-client/cmds/helpers"
-	"github.com/go-go-golems/go-go-mcp/cmd/mcp-client/cmds/layers"
+	"github.com/go-go-golems/go-go-mcp/cmd/go-go-mcp/cmds/client/helpers"
+	"github.com/go-go-golems/go-go-mcp/cmd/go-go-mcp/cmds/client/layers"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
