+++ cmd/go-go-mcp/cmds/client/version.go
