+++ cmd/go-go-mcp/cmds/schema.go
